+++ src/components/Blog/Post/Body/Blockquote.scss
@@ -1,6 +1,6 @@
 @import "common/breakpoints";
 
-@import "./variables";
+@import "./Variables";
 
 blockquote {
   position: relative;
