+++ src/components/Blog/Post/Body/Heading.scss
@@ -1,4 +1,4 @@
-@import "./variables";
+@import "./Variables";
 
 h1 {
   display: none;
