+++ src/components/Blog/Post/Body/ShareLinks.module.scss
@@ -1,6 +1,6 @@
 @import "common/breakpoints";
 
-@import "./variables";
+@import "./Variables";
 
 $breakpoint: 800;
