+++ src/components/Blog/Post/BodyWrapper.module.scss
@@ -1,6 +1,6 @@
 @import "common/breakpoints";
 
-@import "./variables";
+@import "./Variables";
 
 .root {
   max-width: $BlogPost-body-maxWidth-mobile;
