+++ src/components/Blog/Title.js
@@ -2,7 +2,7 @@ import React from "react"
 import PropTypes from "prop-types"
 import classNames from "classnames"
 
-import styles from "./title.module.css"
+import styles from "./Title.module.css"
 
 const BlogTitle = ({ children, className }) => (
   <h1 className={classNames(styles.root, className)}>{children}</h1>
