+++ src/components/Home/VenturesSection/HelpingWithSubsection.js
@@ -2,7 +2,7 @@ import React from "react"
 
 import Text from "src/components/text"
 
-import styles from "./helping_with_subsection.module.css"
+import styles from "./HelpingWithSubsection.module.css"
 
 const HelpingWith = () => (
   <ul className={styles.root}>
