+++ src/components/Logo.js
@@ -4,7 +4,7 @@ import { Link } from "gatsby"
 import classnames from "classnames"
 import _uniqueId from "lodash/uniqueId"
 
-import styles from "./logo.module.css"
+import styles from "./Logo.module.css"
 
 const Logo = ({ color }) => {
   const className = classnames(styles.root, styles[color])
