+++ src/components/SocialLink.js
@@ -1,7 +1,7 @@
 import React from "react"
 import PropTypes from "prop-types"
 
-import Link from "./link"
+import Link from "./Link"
 
 const socialPlatforms = {
   behance: {
