+++ src/components/SubsectionTitle.js
@@ -2,7 +2,7 @@ import React from "react"
 import PropTypes from "prop-types"
 import classNames from "classnames"
 
-import styles from "./subsection_title.module.css"
+import styles from "./SubsectionTitle.module.css"
 
 const SubsectionTitle = ({ children, color }) => {
   const className = classNames(styles.root, styles[color])
